backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
aniso8601==10.0.1
    # via flask-restx
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via flask
certifi==2026.1.4
    # via
    #   backend
    #   httpcore
    #   httpx
    #   pinecone
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via flask
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   pytest
    #   tqdm
cryptography==46.0.5
    # via pyjwt
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via pymongo
flask==3.1.3
    # via
    #   backend
    #   flask-cors
    #   flask-restx
    #   pytest-flask
flask-cors==6.0.2
    # via backend
flask-restx==1.3.2
    # via backend
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.3.2 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
    # via sqlalchemy
gunicorn==25.1.0
    # via backend
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   ollama
    #   openai
httpx-sse==0.4.3
    # via langchain-community
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-resources==6.5.2
    # via flask-restx
iniconfig==2.3.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jiter==0.13.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.26.0
    # via flask-restx
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain==1.2.10
    # via
    #   backend
    #   langchain-mongodb
langchain-classic==1.0.1
    # via
    #   langchain-community
    #   langchain-mongodb
langchain-community==0.4.1
    # via backend
langchain-core==1.2.15
    # via
    #   langchain
    #   langchain-classic
    #   langchain-community
    #   langchain-mongodb
    #   langchain-ollama
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-mongodb==0.11.0
    # via langgraph-checkpoint-mongodb
langchain-ollama==1.0.1
    # via backend
langchain-openai==1.1.10
    # via backend
langchain-text-splitters==1.1.1
    # via
    #   langchain-classic
    #   langchain-mongodb
langgraph==1.0.9
    # via langchain
langgraph-checkpoint==4.0.0
    # via
    #   langgraph
    #   langgraph-checkpoint-mongodb
    #   langgraph-prebuilt
langgraph-checkpoint-mongodb==0.3.1
    # via backend
langgraph-prebuilt==1.0.8
    # via langgraph
langgraph-sdk==0.3.8
    # via langgraph
langsmith==0.7.6
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
lark==1.3.1
    # via langchain-mongodb
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
marshmallow==3.26.2
    # via dataclasses-json
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.4.3
    # via
    #   backend
    #   langchain-community
    #   langchain-mongodb
    #   pandas
ollama==0.6.1
    # via langchain-ollama
openai==2.22.0
    # via langchain-openai
orjson==3.11.7
    # via
    #   langgraph-sdk
    #   langsmith
    #   pinecone
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==24.2
    # via
    #   gunicorn
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   pinecone-plugin-assistant
    #   pytest
pandas==3.0.1
    # via backend
pinecone==8.1.0
    # via backend
pinecone-plugin-assistant==3.0.2
    # via pinecone
pinecone-plugin-interface==0.0.7
    # via pinecone
pluggy==1.6.0
    # via pytest
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   backend
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   ollama
    #   openai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via langchain-community
pygments==2.19.2
    # via pytest
pyjwt==2.11.0
    # via backend
pymongo==4.15.5
    # via
    #   backend
    #   langchain-mongodb
    #   langgraph-checkpoint-mongodb
    #   pymongo-search-utils
pymongo-search-utils==0.3.0
    # via langchain-mongodb
pytest==9.0.2
    # via
    #   backend
    #   pytest-flask
pytest-flask==1.3.0
    # via backend
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   pinecone
python-dotenv==1.2.1
    # via
    #   backend
    #   pydantic-settings
pyyaml==6.0.3
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
referencing==0.37.0
    # via
    #   flask-restx
    #   jsonschema
    #   jsonschema-specifications
regex==2026.2.19
    # via tiktoken
requests==2.32.5
    # via
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   pinecone-plugin-assistant
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via openai
sqlalchemy==2.0.48
    # via
    #   langchain-classic
    #   langchain-community
tenacity==9.1.4
    # via
    #   langchain-community
    #   langchain-core
tiktoken==0.12.0
    # via langchain-openai
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   langchain-core
    #   openai
    #   pinecone
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.3 ; sys_platform == 'emscripten' or sys_platform == 'win32'
    # via pandas
urllib3==2.6.3
    # via
    #   pinecone
    #   requests
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
werkzeug==3.1.6
    # via
    #   flask
    #   flask-cors
    #   flask-restx
    #   pytest-flask
xxhash==3.6.0
    # via
    #   langgraph
    #   langsmith
yarl==1.23.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
